third_party/blink/web_tests/platform/win/external/wpt/css/css-shapes/shape-outside/values/shape-outside-computed-shape-000-expected.txt

This is a testharness.js-based test.
[PASS] em units
[FAIL] ex units
  assert_equals: expected "polygon(130.078px 130.078px)" but got "polygon(130.08px 130.08px)"
[PASS] ch units
[PASS] rem units
Harness: the test ran to completion.
